+++ fuc/api/common.py
@@ -1,7 +1,7 @@
 """
-The ``common`` submodule is used by other ``fuc`` submodules such as
-``pyvcf`` and ``pybed``. It also provides many day-to-day actions used in
-the field of bioinformatics.
+The common submodule is used by other fuc submodules such as pyvcf and
+pybed. It also provides many day-to-day actions used in the field of
+bioinformatics.
 """
 
 import pathlib
@@ -31,7 +31,7 @@ def get_most_similar(a, l):
     i = [i for i, x in enumerate(s) if x == m][0]
     return l[i]
 
-def sumstat(tp, fp, fn, tn):
+def sumstat(fp, fn, tp, tn):
     """Return various important statistics."""
     d = {
         'tpr': tp / (tp + fn), # sensitivity, recall, hit rate