+++ CHANGELOG.rst
@@ -1,14 +1,23 @@
 Changelog
 *********
 
+0.3.0 (2021-04-30)
+------------------
+
+* Rename ``snpeff`` submodule to ``pysnpeff``.
+* Add new submodule ``pyvep``.
+* Update ``fuc.api.pyvcf.VcfFrame`` class.
+* Add ``autodocsumm`` extension to Read the Docs.
+* Add contents to Read the Docs.
+
 0.2.0 (2021-04-26)
 ------------------
 
 * :issue:`2`: Fix Read the Docs automodule not working properly.
 * :issue:`3`: Add ``sphinx-issues`` extension to Read the Docs.
 * Rename submodules ``fuc.api.BedFrame``, ``fuc.api.FastqFrame``, and ``fuc.api.VcfFrame`` to ``fuc.api.pybed``, ``fuc.api.pyfq``, and ``fuc.api.pyvcf``, respectively.
-* Add new methods to ``fuc.api.pyvcf``.
-* Add new methods to ``fuc.api.pyvcf.VcfFrame``.
+* Add new methods to ``fuc.api.pyvcf`` submodule.
+* Add new methods to ``fuc.api.pyvcf.VcfFrame`` class.
 * Add new submodule ``fuc.api.snpeff``.
 
 0.1.4 (2021-04-21)